refactor(hfloat): split hfloat into core / manipulators / iostream (#1334)
Third type of Phase 2 group 5a (#1444).

    core.hpp        59,518 lines / 304 headers / 0 I/O-family
    hfloat.hpp      77,199 lines / 385 headers / 5 I/O-family  (was 78,566 / 399 / 5)

Zero I/O-family headers is met. Under-45,000 lines is not, and reporting the
number rather than moving the bar: utility/decimal_to_binary.hpp, which parse()
converts through, is 57,369 lines on its own (already 0 I/O), so this core is
within 2,200 lines of its single largest dependency. Same shape as dfloat
(#1446); the ceiling belongs to the substrate.

str() had to be reimplemented, and that is this PR's one real change

hfloat's str() is core surface -- operator<< formats through it -- but unlike
dfloat's, which concatenates, it built its result with a stringstream:

    std::stringstream ss;
    if (nrDigits > 0) ss << std::setprecision(nrDigits);
    ss << d;

That single function was the only reason this core would have needed <sstream>.
std::ostream's `defaultfloat` presentation of a double IS printf's %g at the
stream's precision (which defaults to 6), so it now reaches the same formatter
through snprintf("%.*g") -- the <cstdio> idiom Phase 0 adopted -- with a
two-pass call so arbitrary precision still works and a clamp on the size_t to
int narrowing.

Verified, not assumed: a standalone harness compared the two formulations over
604,742 (value, precision) pairs -- the full double exponent range from 1e-320
to 1e308, 40 mantissas per decade, precisions 0 to 60, plus the boundary values
-- with ZERO mismatches on gcc and clang. The differential below then compares
184,176 further str() renderings in situ.

One observable difference remains and is deliberate: a stringstream carries the
global locale while snprintf uses the C locale, so a program that calls
std::locale::global with a non-C numeric facet would previously have seen a
localised decimal separator here and now will not. Nothing in the tree sets one.

What else moved:

- to_binary(), to_hex() and to_native() format through a stringstream ->
  manipulators.hpp.
- operator<< and operator>> -> iostream.hpp, which includes only core.hpp:
  operator<< formats through str() and operator>> hands its token to parse(),
  both core, so there is no dependency on manipulators at all.
- parse() STAYS in the core: it converts through utility/decimal_to_binary.hpp,
  which is already at 0 I/O, so it is arithmetic surface. Same call dfloat's
  parse() got, opposite of cfloat's and bfloat16's istringstream ones.
- native/ieee754.hpp -> native/ieee754_core.hpp.
- HFLOAT_ENABLE_LITERALS, HFLOAT_THROW_ARITHMETIC_EXCEPTION and HFLOAT_EXCEPT
  now default in hfloat_impl.hpp beside the code they govern (#1436).
- manipulators.hpp and attributes.hpp had no includes at all and reached
  std::stringstream, std::setw, Color, type_tag and operator<< entirely through
  the umbrella's include order. Both now name what they use (#1389).

Verification:

- Layer gate: core.hpp alone, and core.hpp + manipulators.hpp, both compile,
  link and run under -Wall -Wpedantic on gcc 13.3 and clang.
- Self-containment: core, manipulators, iostream, attributes and the umbrella
  each compile standalone with 0 errors on both compilers.
- 12 targets built and run green on both: the 10 hfloat tests plus ucalc and
  ucalc_regression.
- Differential against main: hfp32 and hfp64 across every exponent code against
  24 fraction patterns, all three formats across 155 decades x 4 mantissas
  through parse(), each row rendered by to_binary (both nibble settings),
  to_hex, to_native, color_print, str() at 23 precisions, operator<< under six
  stream states, negation, square and double. 7,896,511 bytes of stdout and all
  30 stderr diagnostics byte-identical -- old vs new on gcc, old vs new on
  clang, AND gcc vs clang.
